requirements.txt
Django==1.9.4
djangorestframework
django-filter
django-countries
python-dateutil
pytz